refactor(wallet): changed balance datatype

+++ src/main/java/id/ac/ui/cs/mysawit/payment/dto/WalletTopUpRequestDTO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class WalletTopUpRequestDTO {
     @NotNull
     @Positive
-    private Long amount;
+    private Double amount;
 }
